Rename `crypto` module to `hashes` and move `Blake3Hash` there
