+++ crates/pallet-subspace/src/lib.rs
@@ -112,12 +112,13 @@ pub mod pallet {
     use sp_std::collections::btree_map::BTreeMap;
     use sp_std::num::NonZeroU32;
     use sp_std::prelude::*;
+    use subspace_core_primitives::hashes::Blake3Hash;
     use subspace_core_primitives::pieces::PieceOffset;
     use subspace_core_primitives::pot::PotCheckpoints;
     use subspace_core_primitives::sectors::SectorIndex;
     use subspace_core_primitives::segments::{HistorySize, SegmentHeader, SegmentIndex};
     use subspace_core_primitives::{
-        Blake3Hash, PublicKey, Randomness, RewardSignature, ScalarBytes, SolutionRange,
+        PublicKey, Randomness, RewardSignature, ScalarBytes, SolutionRange,
     };
 
     pub(super) struct InitialSolutionRanges<T: Config> {
