+++ crates/sc-consensus-subspace-rpc/src/lib.rs
@@ -53,10 +53,11 @@ use std::sync::atomic::{AtomicU64, Ordering};
 use std::sync::{Arc, Weak};
 use std::time::Duration;
 use subspace_archiving::archiver::NewArchivedSegment;
+use subspace_core_primitives::hashes::Blake3Hash;
 use subspace_core_primitives::objects::GlobalObjectMapping;
 use subspace_core_primitives::pieces::{Piece, PieceIndex};
 use subspace_core_primitives::segments::{HistorySize, SegmentHeader, SegmentIndex};
-use subspace_core_primitives::{Blake3Hash, BlockHash, PublicKey, SlotNumber, Solution};
+use subspace_core_primitives::{BlockHash, PublicKey, SlotNumber, Solution};
 use subspace_erasure_coding::ErasureCoding;
 use subspace_farmer_components::FarmerProtocolInfo;
 use subspace_kzg::Kzg;
