+++ crates/subspace-core-primitives/src/checksum.rs
@@ -3,7 +3,7 @@
 #[cfg(test)]
 mod tests;
 
-use crate::Blake3Hash;
+use crate::hashes::Blake3Hash;
 use parity_scale_codec::{Decode, Encode, EncodeLike, Error, Input, Output};
 
 /// Output wrapper for SCALE codec that will write Blake3 checksum at the end of the encoding
