+++ crates/subspace-core-primitives/src/checksum/tests.rs
@@ -1,5 +1,5 @@
 use super::Blake3Checksummed;
-use crate::Blake3Hash;
+use crate::hashes::Blake3Hash;
 use parity_scale_codec::{Decode, Encode};
 use rand::prelude::*;
