+++ crates/subspace-core-primitives/src/pot.rs
@@ -1,7 +1,7 @@
 //! Proof of time-related data structures.
 
-use crate::crypto::{blake3_hash, blake3_hash_list};
-use crate::{Blake3Hash, Randomness};
+use crate::hashes::{blake3_hash, blake3_hash_list, Blake3Hash};
+use crate::Randomness;
 use core::fmt;
 use core::num::NonZeroU8;
 use core::str::FromStr;
