+++ crates/subspace-core-primitives/src/segments.rs
@@ -3,9 +3,9 @@
 #[cfg(not(feature = "std"))]
 extern crate alloc;
 
-use crate::crypto::blake3_hash;
+use crate::hashes::{blake3_hash, Blake3Hash};
 use crate::pieces::{FlatPieces, Piece, PieceIndex, RawRecord};
-use crate::{Blake3Hash, BlockNumber};
+use crate::BlockNumber;
 #[cfg(not(feature = "std"))]
 use alloc::boxed::Box;
 use core::array::TryFromSliceError;
