+++ crates/subspace-farmer-components/benches/auditing.rs
@@ -7,10 +7,11 @@ use std::io::Write;
 use std::num::{NonZeroU64, NonZeroUsize};
 use std::{env, fs, slice};
 use subspace_archiving::archiver::Archiver;
+use subspace_core_primitives::hashes::Blake3Hash;
 use subspace_core_primitives::pieces::Record;
 use subspace_core_primitives::sectors::SectorId;
 use subspace_core_primitives::segments::{HistorySize, RecordedHistorySegment};
-use subspace_core_primitives::{Blake3Hash, PublicKey, SolutionRange};
+use subspace_core_primitives::{PublicKey, SolutionRange};
 use subspace_erasure_coding::ErasureCoding;
 use subspace_farmer_components::auditing::audit_plot_sync;
 use subspace_farmer_components::file_ext::{FileExt, OpenOptionsExt};
