+++ crates/subspace-farmer-components/src/auditing.rs
@@ -10,8 +10,9 @@ use crate::{ReadAtOffset, ReadAtSync};
 use rayon::prelude::*;
 use std::collections::HashSet;
 use std::io;
+use subspace_core_primitives::hashes::Blake3Hash;
 use subspace_core_primitives::sectors::{SBucket, SectorId, SectorIndex, SectorSlotChallenge};
-use subspace_core_primitives::{Blake3Hash, PublicKey, ScalarBytes, SolutionRange};
+use subspace_core_primitives::{PublicKey, ScalarBytes, SolutionRange};
 use subspace_verification::is_within_solution_range;
 use thiserror::Error;
