+++ crates/subspace-farmer-components/src/plotting.rs
@@ -24,12 +24,12 @@ use std::simd::Simd;
 use std::sync::atomic::{AtomicBool, Ordering};
 use std::sync::Arc;
 use std::time::Duration;
-use subspace_core_primitives::crypto::{blake3_hash, blake3_hash_parallel};
+use subspace_core_primitives::hashes::{blake3_hash, blake3_hash_parallel, Blake3Hash};
 use subspace_core_primitives::pieces::{PieceIndex, PieceOffset, Record};
 use subspace_core_primitives::pos::PosSeed;
 use subspace_core_primitives::sectors::{SBucket, SectorId, SectorIndex};
 use subspace_core_primitives::segments::HistorySize;
-use subspace_core_primitives::{Blake3Hash, PublicKey, ScalarBytes};
+use subspace_core_primitives::{PublicKey, ScalarBytes};
 use subspace_erasure_coding::ErasureCoding;
 use subspace_kzg::{Kzg, Scalar};
 use subspace_proof_of_space::{Table, TableGenerator};
