+++ crates/subspace-farmer/src/bin/subspace-farmer/commands/benchmark.rs
@@ -8,8 +8,9 @@ use std::collections::HashSet;
 use std::fs::OpenOptions;
 use std::num::NonZeroUsize;
 use std::path::PathBuf;
+use subspace_core_primitives::hashes::Blake3Hash;
 use subspace_core_primitives::pieces::Record;
-use subspace_core_primitives::{Blake3Hash, SolutionRange};
+use subspace_core_primitives::SolutionRange;
 use subspace_erasure_coding::ErasureCoding;
 use subspace_farmer::single_disk_farm::direct_io_file::DirectIoFile;
 use subspace_farmer::single_disk_farm::farming::rayon_files::RayonFiles;
