+++ crates/subspace-farmer/src/bin/subspace-farmer/commands/cluster/controller/farms.rs
@@ -19,8 +19,8 @@ use std::mem;
 use std::pin::{pin, Pin};
 use std::sync::Arc;
 use std::time::Instant;
+use subspace_core_primitives::hashes::Blake3Hash;
 use subspace_core_primitives::sectors::SectorIndex;
-use subspace_core_primitives::Blake3Hash;
 use subspace_farmer::cluster::controller::ClusterControllerFarmerIdentifyBroadcast;
 use subspace_farmer::cluster::farmer::{ClusterFarm, ClusterFarmerIdentifyFarmBroadcast};
 use subspace_farmer::cluster::nats_client::NatsClient;
