+++ crates/subspace-farmer/src/cluster/farmer.rs
@@ -27,10 +27,9 @@ use std::future::{pending, Future};
 use std::pin::{pin, Pin};
 use std::sync::Arc;
 use std::time::{Duration, Instant};
-use subspace_core_primitives::crypto::blake3_hash_list;
+use subspace_core_primitives::hashes::{blake3_hash_list, Blake3Hash};
 use subspace_core_primitives::pieces::{Piece, PieceOffset};
 use subspace_core_primitives::sectors::SectorIndex;
-use subspace_core_primitives::Blake3Hash;
 use subspace_farmer_components::plotting::PlottedSector;
 use subspace_rpc_primitives::SolutionResponse;
 use tokio::time::MissedTickBehavior;
