+++ crates/subspace-farmer/src/disk_piece_cache.rs
@@ -19,9 +19,8 @@ use std::path::Path;
 use std::sync::Arc;
 use std::task::Poll;
 use std::{fs, io};
-use subspace_core_primitives::crypto::blake3_hash_list;
+use subspace_core_primitives::hashes::{blake3_hash_list, Blake3Hash};
 use subspace_core_primitives::pieces::{Piece, PieceIndex};
-use subspace_core_primitives::Blake3Hash;
 use subspace_farmer_components::file_ext::FileExt;
 use thiserror::Error;
 use tokio::runtime::Handle;
