+++ crates/subspace-networking/src/behavior/persistent_parameters.rs
@@ -19,8 +19,7 @@ use std::str::FromStr;
 use std::sync::Arc;
 use std::time::{Duration, SystemTime};
 use std::{io, mem};
-use subspace_core_primitives::crypto::blake3_hash;
-use subspace_core_primitives::Blake3Hash;
+use subspace_core_primitives::hashes::{blake3_hash, Blake3Hash};
 use thiserror::Error;
 use tokio::time::{sleep, Sleep};
 use tracing::{debug, error, trace, warn};
