+++ crates/subspace-proof-of-space/src/chiapos/table.rs
@@ -26,7 +26,7 @@ use seq_macro::seq;
 #[cfg(all(not(feature = "std"), any(feature = "parallel", test)))]
 use spin::Mutex;
 use static_assertions::const_assert;
-use subspace_core_primitives::crypto::{blake3_hash, blake3_hash_list};
+use subspace_core_primitives::hashes::{blake3_hash, blake3_hash_list};
 
 pub(super) const COMPUTE_F1_SIMD_FACTOR: usize = 8;
 pub(super) const FIND_MATCHES_AND_COMPUTE_UNROLL_FACTOR: usize = 8;
