+++ crates/subspace-proof-of-space/src/shim.rs
@@ -3,7 +3,7 @@
 
 use crate::{PosTableType, Table, TableGenerator};
 use core::iter;
-use subspace_core_primitives::crypto::blake3_hash;
+use subspace_core_primitives::hashes::blake3_hash;
 use subspace_core_primitives::pos::{PosProof, PosSeed};
 use subspace_core_primitives::U256;
